push

+++ 07_pydanticai/Kokchun/b_pydanticai_fastapi/setup_db.py
@@ -5,6 +5,6 @@ query_duckdb("""
         title TEXT,
         year INTEGER,
         genre TEXT,
-        rating TINYINT
+        rating INTEGER,
     );
     """)
